Move validations to dbt_semantic_interfaces
