+++ dbt_semantic_interfaces/dbt_semantic_interfaces/validations/common_entities.py
@@ -4,7 +4,7 @@ from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
 from dbt_semantic_interfaces.objects.elements.entity import Entity
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import SemanticModelElementReference, EntityReference
-from metricflow.model.validations.validator_helpers import (
+from dbt_semantic_interfaces.validations.validator_helpers import (
     SemanticModelElementContext,
     SemanticModelElementType,
     FileContext,
