+++ dbt_semantic_interfaces/dbt_semantic_interfaces/validations/element_const.py
@@ -3,7 +3,7 @@ from typing import List, DefaultDict
 
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import SemanticModelReference
-from metricflow.model.validations.validator_helpers import (
+from dbt_semantic_interfaces.validations.validator_helpers import (
     SemanticModelContext,
     SemanticModelElementType,
     FileContext,
