+++ dbt_semantic_interfaces/dbt_semantic_interfaces/validations/entities.py
@@ -6,7 +6,7 @@ from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
 from dbt_semantic_interfaces.objects.elements.entity import EntityType
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import SemanticModelReference
-from metricflow.model.validations.validator_helpers import (
+from dbt_semantic_interfaces.validations.validator_helpers import (
     SemanticModelContext,
     FileContext,
     ModelValidationRule,
@@ -14,7 +14,7 @@ from metricflow.model.validations.validator_helpers import (
     ValidationError,
     validate_safely,
 )
-from metricflow.model.validations.validator_helpers import ValidationFutureError
+from dbt_semantic_interfaces.validations.validator_helpers import ValidationFutureError
 
 logger = logging.getLogger(__name__)
