+++ dbt_semantic_interfaces/dbt_semantic_interfaces/validations/measures.py
@@ -8,8 +8,8 @@ from dbt_semantic_interfaces.objects.elements.dimension import DimensionType
 from dbt_semantic_interfaces.objects.metric import Metric
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import MetricModelReference, MeasureReference
-from metricflow.model.validations.unique_valid_name import UniqueAndValidNameRule
-from metricflow.model.validations.validator_helpers import (
+from dbt_semantic_interfaces.validations.unique_valid_name import UniqueAndValidNameRule
+from dbt_semantic_interfaces.validations.validator_helpers import (
     SemanticModelElementContext,
     SemanticModelElementReference,
     SemanticModelElementType,
