+++ dbt_semantic_interfaces/dbt_semantic_interfaces/validations/metrics.py
@@ -5,8 +5,8 @@ from dbt_semantic_interfaces.errors import ParsingException
 from dbt_semantic_interfaces.objects.metric import Metric, MetricType, MetricTimeWindow
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import MetricModelReference
-from metricflow.model.validations.unique_valid_name import UniqueAndValidNameRule
-from metricflow.model.validations.validator_helpers import (
+from dbt_semantic_interfaces.validations.unique_valid_name import UniqueAndValidNameRule
+from dbt_semantic_interfaces.validations.validator_helpers import (
     FileContext,
     MetricContext,
     ModelValidationRule,
