+++ dbt_semantic_interfaces/dbt_semantic_interfaces/validations/non_empty.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
-from metricflow.model.validations.validator_helpers import (
+from dbt_semantic_interfaces.validations.validator_helpers import (
     ModelValidationRule,
     ValidationError,
     ValidationIssue,
