+++ metricflow/dataset/dataset.py
@@ -5,7 +5,7 @@ from typing import Sequence
 
 from dbt_semantic_interfaces.references import TimeDimensionReference
 from metricflow.instances import InstanceSet, TimeDimensionInstance
-from metricflow.model.validations.unique_valid_name import MetricFlowReservedKeywords
+from dbt_semantic_interfaces.validations.unique_valid_name import MetricFlowReservedKeywords
 from metricflow.specs import TimeDimensionSpec
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
