+++ metricflow/model/dbt_converter.py
@@ -37,7 +37,11 @@ from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
 from dbt_semantic_interfaces.objects.metric import Metric
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.parsing.dir_to_model import ModelBuildResult
-from metricflow.model.validations.validator_helpers import ModelValidationResults, ValidationError, ValidationIssue
+from dbt_semantic_interfaces.validations.validator_helpers import (
+    ModelValidationResults,
+    ValidationError,
+    ValidationIssue,
+)
 
 logger = logging.getLogger(__name__)
