+++ metricflow/model/dbt_mapping_rules/dbt_mapping_rule.py
@@ -6,7 +6,7 @@ from typing_extensions import TypeAlias
 
 from dbt_metadata_client.dbt_metadata_api_schema import MetricNode, ModelNode
 from dbt_semantic_interfaces.objects.metric import MetricType
-from metricflow.model.validations.validator_helpers import ModelValidationResults
+from dbt_semantic_interfaces.validations.validator_helpers import ModelValidationResults
 
 
 TransformedObjectsValueType: TypeAlias = Any  # type: ignore[misc]
