+++ metricflow/model/transformations/dbt_to_metricflow.py
@@ -19,7 +19,11 @@ from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.parsing.dir_to_model import ModelBuildResult
-from metricflow.model.validations.validator_helpers import ModelValidationResults, ValidationError, ValidationIssue
+from dbt_semantic_interfaces.validations.validator_helpers import (
+    ModelValidationResults,
+    ValidationError,
+    ValidationIssue,
+)
 
 
 @dataclass
