+++ metricflow/test/api/test_metricflow_client.py
@@ -1,7 +1,7 @@
 from metricflow.api.metricflow_client import MetricFlowClient
 from metricflow.dataflow.sql_table import SqlTable
 from metricflow.engine.models import Dimension, Metric
-from metricflow.model.validations.validator_helpers import ModelValidationResults
+from dbt_semantic_interfaces.validations.validator_helpers import ModelValidationResults
 from metricflow.random_id import random_id
