+++ metricflow/test/model/parsing/test_metric_parsing.py
@@ -5,7 +5,7 @@ from dbt_semantic_interfaces.objects.metric import MetricTimeWindow, MetricInput
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.parsing.dir_to_model import parse_yaml_files_to_model
 from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
-from metricflow.model.validations.validator_helpers import ModelValidationException
+from dbt_semantic_interfaces.validations.validator_helpers import ModelValidationException
 
 
 def test_legacy_measure_metric_parsing() -> None:
