+++ metricflow/test/model/validations/test_common_entities.py
@@ -3,10 +3,10 @@ import pytest
 import re
 from typing import Callable
 
-from metricflow.model.model_validator import ModelValidator
+from dbt_semantic_interfaces.model_validator import ModelValidator
 from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
-from metricflow.model.validations.common_entities import CommonEntitysRule
+from dbt_semantic_interfaces.validations.common_entities import CommonEntitysRule
 from metricflow.specs import EntitySpec
 from metricflow.test.test_utils import find_semantic_model_with
