+++ metricflow/test/model/validations/test_element_const.py
@@ -2,12 +2,12 @@ import copy
 import pytest
 from typing import Tuple
 
-from metricflow.model.model_validator import ModelValidator
+from dbt_semantic_interfaces.model_validator import ModelValidator
 from dbt_semantic_interfaces.objects.semantic_model import SemanticModel
 from dbt_semantic_interfaces.objects.elements.dimension import Dimension, DimensionType
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
-from metricflow.model.validations.element_const import ElementConsistencyRule
-from metricflow.model.validations.validator_helpers import SemanticModelElementType, ModelValidationException
+from dbt_semantic_interfaces.validations.element_const import ElementConsistencyRule
+from dbt_semantic_interfaces.validations.validator_helpers import SemanticModelElementType, ModelValidationException
 from metricflow.test.test_utils import find_semantic_model_with
