+++ metricflow/test/model/validations/test_reserved_keywords.py
@@ -2,8 +2,8 @@ import copy
 import random
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.objects.semantic_model import NodeRelation
-from metricflow.model.validations.reserved_keywords import RESERVED_KEYWORDS, ReservedKeywordsRule
-from metricflow.model.validations.validator_helpers import ValidationIssueLevel
+from dbt_semantic_interfaces.validations.reserved_keywords import RESERVED_KEYWORDS, ReservedKeywordsRule
+from dbt_semantic_interfaces.validations.validator_helpers import ValidationIssueLevel
 from metricflow.test.test_utils import find_semantic_model_with
