+++ metricflow/test/model/validations/test_semantic_models.py
@@ -1,7 +1,7 @@
 import pytest
 
 from dbt_semantic_interfaces.objects.elements.dimension import Dimension, DimensionType, DimensionTypeParams
-from metricflow.model.validations.validator_helpers import ModelValidationException
+from dbt_semantic_interfaces.validations.validator_helpers import ModelValidationException
 from metricflow.test.model.validations.helpers import semantic_model_with_guaranteed_meta
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
