+++ metricflow/test/model/validations/test_validator_helpers.py
@@ -8,7 +8,7 @@ from dbt_semantic_interfaces.references import (
     MetricModelReference,
 )
 
-from metricflow.model.validations.validator_helpers import (
+from dbt_semantic_interfaces.validations.validator_helpers import (
     SemanticModelContext,
     SemanticModelElementContext,
     SemanticModelElementType,
